admin can delete any tweet from now

+++ mainapp/classes/AjaxHandle.py
@@ -644,7 +644,8 @@ class AjaxHandle(AjaxSearch):
             if task == 'delete':
             	'''Deleting a Tweet'''
                 #print change_id
-                tweet_feed_obj.delete_tweet(request.user.id,str(change_id))
+                if request.user.is_superuser or request.user.is_authenticated:
+                    tweet_feed_obj.delete_tweet(request.user.id,str(change_id))
                 return HttpResponse(json.dumps({'status':'1', 'activity':'deleteTweet', '_id':change_id}))
 
             if task == 'follow':
